# expression: implement vectorized evaluation for `JSONArrayAppend` 