<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #12104 <!-- REMOVE this line if no issue to close -->

Problem Summary: implement vectorized evaluation for `JSONArrayAppend`

### What is changed and how it works?
implement vectorized evaluation for `JSONArrayAppend`

```
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONArrayAppendSig-VecBuiltinFunc-12               868           1280255 ns/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONArrayAppendSig-NonVecBuiltinFunc-12            920           1284265 ns/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONArrayAppendSig-VecBuiltinFunc#01-12            481           2462974 ns/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONArrayAppendSig-NonVecBuiltinFunc#01-12         470           2485364 ns/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONArrayAppendSig-VecBuiltinFunc#02-12            692           1786525 ns/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONArrayAppendSig-NonVecBuiltinFunc#02-12         680           1751424 ns/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test