@zz-jason 2020-04-14
/run-sqllogic-test-2